+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/utils.ts
@@ -4,6 +4,7 @@ import CommandFactory from 'sap/ui/rta/command/CommandFactory';
 
 import { QuickActionContext } from '../../../cpe/quick-actions/quick-action-definition';
 import { getUi5Version, isLowerThanMinimalUi5Version, isVersionEqualOrHasNewerPatch } from '../../../utils/version';
+import { Manifest } from 'sap/ui/rta/RuntimeAuthoring';
 
 /**
  * Prepares the change for the manifest setting.
@@ -53,20 +54,29 @@ export async function prepareManifestChange(
 }
 
 /**
- * Checks if the current UI5 version supports manifest changes in v2 applications.
+ * Checks if the current UI5 version and manifest structure is supported in v2 applications.
  *
- * Returns `true` if the UI5 version is supported.
- * Otherwise, returns `false`.
+ * @param manifest - manifest changes of the current application.
+ *
+ * Returns `false`
+ *
+ *  - If the manifest is structured is an array
+ *  - If the UI5 version is not supported
+ * Otherwise, returns `true`.
  *
  */
-export async function areManifestChangesSupported(): Promise<boolean> {
+export async function areManifestChangesSupported(manifest: Manifest): Promise<boolean> {
+    const pagesStructureInManifest = manifest['sap.ui.generic.app'].pages;
+    if (Array.isArray(pagesStructureInManifest)) {
+        return false;
+    }
+
     const version = await getUi5Version();
-    return (
-        isLowerThanMinimalUi5Version(version, { major: 1, minor: 128 }) &&
-        !(
-            isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 96, patch: 37 }) ||
-            isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 108, patch: 38 }) ||
-            isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 120, patch: 23 })
-        )
-    );
+    const isAboveOrEqualMinimalVersion = !isLowerThanMinimalUi5Version(version, { major: 1, minor: 128 });
+    const isSupportedPatchVersion =
+        isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 96, patch: 37 }) ||
+        isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 108, patch: 38 }) ||
+        isVersionEqualOrHasNewerPatch(version, { major: 1, minor: 120, patch: 23 });
+
+    return isAboveOrEqualMinimalVersion || isSupportedPatchVersion;
 }