+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.7.13
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+    -   @sap-ux/telemetry@0.5.47
+
 ## 0.7.12
 
 ### Patch Changes